pyproject.toml
[project]
name = "Convertly"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [{name = "Ditpowuh"}]
dependencies = [
  "flet==0.28.3",
  "pydub==0.25.1",
  "moviepy==2.2.1",
  "fonttools==4.59.2",
  "pillow==11.3.0",
  "pillow_heif==1.1.0",
  "pdf2image==1.17.0",
  "opencv-python==4.12.0.88",
  "numpy==2.2.6"
]

[tool.flet]
product = "Convertly"
company = "Ditpowuh"
copyright = "Copyright (C) Ditpowuh 2025"

[tool.flet.app]
path = "src"

[tool.flet.app.boot_screen]
show = true
message = "Preparing Convertly..."

[tool.flet.app.startup_screen]
show = true
message = "Starting up Convertly..."

[tool.uv]
dev-dependencies = [
  "flet[all]==0.28.3"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.3"}
